fix: resolve query error caused by float types

+++ packages/andromeda-math/src/curve.rs
@@ -60,7 +60,7 @@ pub enum QueryMsg {
     #[returns(GetCurveConfigResponse)]
     GetCurveConfig {},
     #[returns(GetPlotYFromXResponse)]
-    GetPlotYFromX { x_value: f64 },
+    GetPlotYFromX { x_value: u64 },
 }
 
 #[cw_serde]
